feat(assessments): edit now linked in routes
