+++ app/views/course/assessment/assessments/edit.html.slim
@@ -1,3 +1,2 @@
 - add_breadcrumb format_inline_text(@assessment.title)
-= page_header
-div#assessment-edit
+div#course-assessments
