+++ client/app/bundles/course/assessment/assessments.js
@@ -1,3 +1,2 @@
 require('./assessments-authenticate');
-require('./assessments-edit');
 require('./assessments-index');
