--- client/app/bundles/course/assessment/pages/AssessmentEdit/translations.intl.js
@@ -1,18 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-const translations = defineMessages({
-  updateAssessment: {
-    id: 'course.assessment.edit.update',
-    defaultMessage: 'Save',
-  },
-  updateSuccess: {
-    id: 'course.assessment.update.success',
-    defaultMessage: 'Assessment was updated.',
-  },
-  updateFailure: {
-    id: 'course.assessment.update.fail',
-    defaultMessage: 'Failed to update assessment.',
-  },
-});
-
-export default translations;
