+++ client/app/bundles/course/assessment/translations.ts
@@ -1,6 +1,18 @@
 import { defineMessages } from 'react-intl';
 
 const translations = defineMessages({
+  updateAssessment: {
+    id: 'course.assessment.edit.update',
+    defaultMessage: 'Save',
+  },
+  updateSuccess: {
+    id: 'course.assessment.update.success',
+    defaultMessage: 'Assessment was updated.',
+  },
+  updateFailure: {
+    id: 'course.assessment.update.fail',
+    defaultMessage: 'Failed to update assessment.',
+  },
   newAssessment: {
     id: 'course.assessment.newAssessment',
     defaultMessage: 'New Assessment',
